Add DDL for metrics
